+++ src/ts/blocks/SampleBlock/index.ts
@@ -1,13 +1,9 @@
-import { registerBlockType } from '@wordpress/blocks';
+import { registerBlockType, type BlockConfiguration } from '@wordpress/blocks';
 import Edit from './edit';
 import Save from './save';
 import metadata from './block.json';
 
-registerBlockType(
-    metadata.name, 
-    {
-        ...metadata, 
-        edit: Edit, 
-        save: Save
-    } as any
-);
+registerBlockType(metadata as BlockConfiguration, {
+    edit: Edit,
+    save: Save,
+});