+++ src/ts/blocks/index.tsx
@@ -1,1 +1,1 @@
-import './SampleBlock'
+import './SampleBlock';