+++ src/ts/classic/index.ts
@@ -1,6 +1,3 @@
-document.addEventListener(
-    'DOMContentLoaded', 
-    () => {
-        console.log('Classic mode loaded.');
-    }
-)
+document.addEventListener('DOMContentLoaded', () => {
+    console.log('Classic mode loaded.');
+});